+++ hcpworks/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to the "hcpworks" extension will be documented in this file.
 ## [0.0.13]
 
 - Preview no longer steals focus when opened from the module list
+- Display `\table` and the diagram in vertically split panes that can be scrolled independently.
+    - Allow the splitter between the table pane and the SVG pane to be dragged to resize their heights.
+    - If there is no table, display only the diagram at full size as before.
+    - Enable zooming with Ctrl+Wheel only in the SVG pane, without interfering with scrolling in the table pane.
 
 ## [0.0.12]